Add relation traits for DefectReport, PurchaseOrder, ReportAudit, User, Vehicle, and Work models
