+++ app/Models/PurchaseOrder.php
@@ -6,11 +6,12 @@ use App\Models\Traits\AttachmentStatus;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\SoftDeletes;
+use App\Models\Traits\PurchaseOrderRelation;
 use Illuminate\Support\Str;
 
 class PurchaseOrder extends Model
 {
-    use AttachmentStatus, HasFactory, SoftDeletes;
+    use AttachmentStatus, HasFactory, SoftDeletes,PurchaseOrderRelation;
 
     protected $fillable = [
         'defect_report_id',
@@ -28,20 +29,6 @@ class PurchaseOrder extends Model
     ];
 
     // Relationships
-    public function defectReport()
-    {
-        return $this->belongsTo(DefectReport::class);
-    }
-
-    public function creator()
-    {
-        return $this->belongsTo(User::class, 'created_by');
-    }
-
-    public function works()
-    {
-        return $this->hasMany(Work::class, 'purchase_order_id');
-    }
 
     // Scopes for role-based filtering
     public function scopeForUser($query, $user)
